+++ src/api/index.js
@@ -1,0 +1,13 @@
+import {apiGetAuth, apiSetAuth} from './auth';
+import {apiGetProducts, apiGetProduct, apiGetProductCategory} from './products';
+import {apiGetCart,apiAddToCart} from './cart';
+
+export {
+    apiGetAuth,
+    apiSetAuth,
+    apiGetProducts,
+    apiGetProduct,
+    apiGetProductCategory,
+    apiGetCart,
+    apiAddToCart,
+};
