+++ src/api/products.js
@@ -1,0 +1,19 @@
+const axios = require('axios');
+
+const apiGetProducts = () => {
+    return axios.get('/api/products');
+};
+
+const apiGetProduct = (productId) => {
+    return axios.get(`/api/products/${productId}`);
+};
+
+const apiGetProductCategory = (categoryId) => {
+    return axios.get(`/api/products/category/${categoryId}`);
+};
+
+export {
+    apiGetProducts,
+    apiGetProduct,
+    apiGetProductCategory,
+};
